+++ simplelink/kernel/zephyr/dpl/stubs.h
@@ -1,0 +1,8 @@
+#ifndef STUBS_H_
+#define STUBS_H_
+
+#include <zephyr.h>
+
+#define STUB(fmt, args...) printk("%s: %s(): %d: STUB: " fmt "\n", __FILE__, __func__, __LINE__, ##args)
+
+#endif /* STUBS_H_ */
